Move value type-checking fn to a `Type` method.
