+++ trustfall_core/src/ir/types/mod.rs
@@ -2,6 +2,6 @@ mod base;
 mod operations;
 
 pub use base::Type;
-pub use operations::{is_argument_type_valid, NamedTypedValue};
+pub use operations::NamedTypedValue;
 
 pub(crate) use operations::{is_base_type_orderable, is_scalar_only_subtype};
