+++ trustfall_core/src/ir/types/operations.rs
@@ -2,8 +2,8 @@ use std::fmt::Debug;
 
 use super::{
     super::{
-        Argument, ContextField, FieldRef, FieldValue, FoldSpecificField, FoldSpecificFieldKind,
-        LocalField, VariableRef,
+        Argument, ContextField, FieldRef, FoldSpecificField, FoldSpecificFieldKind, LocalField,
+        VariableRef,
     },
     Type,
 };
@@ -130,251 +130,5 @@ pub(crate) fn is_scalar_only_subtype(parent_type: &Type, maybe_subtype: &Type) -
         (None, None) => true,
         (Some(parent), Some(maybe_subtype)) => is_scalar_only_subtype(&parent, &maybe_subtype),
         _ => false,
-    }
-}
-
-/// Check if the given argument value is valid for the specified variable type.
-///
-/// In particular, mixed integer types in a list are considered valid for types like `[Int]`.
-/// ```rust
-/// use trustfall_core::ir::{FieldValue, types::{Type, is_argument_type_valid}};
-///
-/// let variable_type = Type::parse("[Int]").unwrap();
-/// let argument_value = FieldValue::List([
-///     FieldValue::Int64(-1),
-///     FieldValue::Uint64(1),
-///     FieldValue::Null,
-/// ].as_slice().into());
-/// assert!(is_argument_type_valid(&variable_type, &argument_value));
-/// ```
-pub fn is_argument_type_valid(variable_type: &Type, argument_value: &FieldValue) -> bool {
-    match argument_value {
-        FieldValue::Null => {
-            // This is a valid value only if this layer is nullable.
-            variable_type.nullable()
-        }
-        FieldValue::Int64(_) | FieldValue::Uint64(_) => {
-            // This is a valid value only if the type is Int, ignoring nullability.
-            !variable_type.is_list() && variable_type.base_type() == "Int"
-        }
-        FieldValue::Float64(_) => {
-            // This is a valid value only if the type is Float, ignoring nullability.
-            !variable_type.is_list() && variable_type.base_type() == "Float"
-        }
-        FieldValue::String(_) => {
-            // This is a valid value only if the type is String, ignoring nullability.
-            !variable_type.is_list() && variable_type.base_type() == "String"
-        }
-        FieldValue::Boolean(_) => {
-            // This is a valid value only if the type is Boolean, ignoring nullability.
-            !variable_type.is_list() && variable_type.base_type() == "Boolean"
-        }
-        FieldValue::List(nested_values) => {
-            // This is a valid value only if the type is a list, and all the inner elements
-            // are valid instances of the type inside the list.
-            if let Some(list) = variable_type.as_list() {
-                nested_values.iter().all(|value| is_argument_type_valid(&list, value))
-            } else {
-                false
-            }
-        }
-        FieldValue::Enum(_) => todo!(),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use itertools::Itertools;
-
-    use crate::ir::{
-        types::{is_argument_type_valid, Type},
-        FieldValue,
-    };
-
-    #[test]
-    fn null_values_are_only_valid_for_nullable_types() {
-        let nullable_types = [
-            Type::parse("Int").unwrap(),
-            Type::parse("String").unwrap(),
-            Type::parse("Boolean").unwrap(),
-            Type::parse("[Int!]").unwrap(),
-            Type::parse("[[Int!]!]").unwrap(),
-        ];
-        let non_nullable_types =
-            nullable_types.iter().map(|t| t.with_nullability(false)).collect_vec();
-
-        for nullable_type in &nullable_types {
-            assert!(is_argument_type_valid(nullable_type, &FieldValue::Null), "{}", nullable_type);
-        }
-        for non_nullable_type in &non_nullable_types {
-            assert!(
-                !is_argument_type_valid(non_nullable_type, &FieldValue::Null),
-                "{}",
-                non_nullable_type
-            );
-        }
-    }
-
-    #[test]
-    fn int_values_are_valid_only_for_int_type_regardless_of_nullability() {
-        let matching_types = [Type::parse("Int").unwrap(), Type::parse("Int!").unwrap()];
-        let non_matching_types = [
-            Type::parse("String").unwrap(),
-            Type::parse("[Int!]").unwrap(),
-            Type::parse("[Int!]!").unwrap(),
-            Type::parse("[[Int!]!]").unwrap(),
-        ];
-        let values = [
-            FieldValue::Int64(-42),
-            FieldValue::Int64(0),
-            FieldValue::Uint64(0),
-            FieldValue::Uint64((i64::MAX as u64) + 1),
-        ];
-
-        for value in &values {
-            for matching_type in &matching_types {
-                assert!(is_argument_type_valid(matching_type, value), "{matching_type} {value:?}",);
-            }
-            for non_matching_type in &non_matching_types {
-                assert!(
-                    !is_argument_type_valid(non_matching_type, value),
-                    "{non_matching_type} {value:?}",
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn string_values_are_valid_only_for_string_type_regardless_of_nullability() {
-        let matching_types = [Type::parse("String").unwrap(), Type::parse("String!").unwrap()];
-        let non_matching_types = [
-            Type::parse("Int").unwrap(),
-            Type::parse("[String!]").unwrap(),
-            Type::parse("[String!]!").unwrap(),
-            Type::parse("[[String!]!]").unwrap(),
-        ];
-        let values = [
-            FieldValue::String("".into()), // empty string is not the same value as null
-            FieldValue::String("test string".into()),
-        ];
-
-        for value in &values {
-            for matching_type in &matching_types {
-                assert!(is_argument_type_valid(matching_type, value), "{matching_type} {value:?}",);
-            }
-            for non_matching_type in &non_matching_types {
-                assert!(
-                    !is_argument_type_valid(non_matching_type, value),
-                    "{non_matching_type} {value:?}",
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn boolean_values_are_valid_only_for_boolean_type_regardless_of_nullability() {
-        let matching_types = [Type::parse("Boolean").unwrap(), Type::parse("Boolean!").unwrap()];
-        let non_matching_types = [
-            Type::parse("Int").unwrap(),
-            Type::parse("[Boolean!]").unwrap(),
-            Type::parse("[Boolean!]!").unwrap(),
-            Type::parse("[[Boolean!]!]").unwrap(),
-        ];
-        let values = [FieldValue::Boolean(false), FieldValue::Boolean(true)];
-
-        for value in &values {
-            for matching_type in &matching_types {
-                assert!(is_argument_type_valid(matching_type, value), "{matching_type} {value:?}",);
-            }
-            for non_matching_type in &non_matching_types {
-                assert!(
-                    !is_argument_type_valid(non_matching_type, value),
-                    "{non_matching_type} {value:?}",
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn list_types_correctly_check_contents_of_list() {
-        let non_nullable_contents_matching_types =
-            vec![Type::parse("[Int!]").unwrap(), Type::parse("[Int!]!").unwrap()];
-        let nullable_contents_matching_types =
-            vec![Type::parse("[Int]").unwrap(), Type::parse("[Int]!").unwrap()];
-        let non_matching_types = [
-            Type::parse("Int").unwrap(),
-            Type::parse("Int!").unwrap(),
-            Type::parse("[String!]").unwrap(),
-            Type::parse("[String!]!").unwrap(),
-            Type::parse("[[String!]!]").unwrap(),
-        ];
-        let non_nullable_values = [
-            FieldValue::List((1..3).map(FieldValue::Int64).collect_vec().into()),
-            FieldValue::List((1..3).map(FieldValue::Uint64).collect_vec().into()),
-            FieldValue::List(
-                vec![
-                    // Integer-typed but non-homogeneous FieldValue entries are okay.
-                    FieldValue::Int64(-42),
-                    FieldValue::Uint64(64),
-                ]
-                .into(),
-            ),
-        ];
-        let nullable_values = [
-            FieldValue::List(
-                vec![FieldValue::Int64(1), FieldValue::Null, FieldValue::Int64(2)].into(),
-            ),
-            FieldValue::List(vec![FieldValue::Null, FieldValue::Uint64(42)].into()),
-            FieldValue::List(
-                vec![
-                    // Integer-typed but non-homogeneous FieldValue entries are okay.
-                    FieldValue::Int64(-1),
-                    FieldValue::Uint64(1),
-                    FieldValue::Null,
-                ]
-                .into(),
-            ),
-        ];
-
-        for value in &non_nullable_values {
-            // Values without nulls match both the nullable and the non-nullable types.
-            for matching_type in &nullable_contents_matching_types {
-                assert!(is_argument_type_valid(matching_type, value), "{matching_type} {value:?}",);
-            }
-            for matching_type in &non_nullable_contents_matching_types {
-                assert!(is_argument_type_valid(matching_type, value), "{matching_type} {value:?}",);
-            }
-
-            // Regardless of nulls, these types don't match.
-            for non_matching_type in &non_matching_types {
-                assert!(
-                    !is_argument_type_valid(non_matching_type, value),
-                    "{non_matching_type} {value:?}",
-                );
-            }
-        }
-
-        for value in &nullable_values {
-            // Nullable values match only the nullable types.
-            for matching_type in &nullable_contents_matching_types {
-                assert!(is_argument_type_valid(matching_type, value), "{matching_type} {value:?}",);
-            }
-
-            // The nullable values don't match the non-nullable types.
-            for non_matching_type in &non_nullable_contents_matching_types {
-                assert!(
-                    !is_argument_type_valid(non_matching_type, value),
-                    "{non_matching_type} {value:?}",
-                );
-            }
-
-            // Regardless of nulls, these types don't match.
-            for non_matching_type in &non_matching_types {
-                assert!(
-                    !is_argument_type_valid(non_matching_type, value),
-                    "{non_matching_type} {value:?}",
-                );
-            }
-        }
     }
 }
